requirements.txt
gunicorn>=23.0.0
flask>=3.1.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.43
psycopg2-binary==2.9.9
Flask-Limiter==3.13
Flask-Caching==2.3.1
Flask-Marshmallow==1.3.0
marshmallow-sqlalchemy==1.4.2
marshmallow==4.0.1
PyJWT==2.10.1
email-validator==2.3.0
bcrypt==4.3.0
Flask-JWT-Extended==4.7.1
Flask-CORS==6.0.1
flasgger==0.9.7.1
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
itsdangerous==2.2.0
click==8.3.0
blinker==1.9.0
greenlet==3.2.4
typing_extensions==4.15.0
bandit>=1.7.5
safety>=2.3.0
watchdog>=3.0.0
python-dotenv>=1.0.0
redis>=5.0.0